+++ examples/basic_examples/amqp_middleware_service.py
@@ -4,12 +4,14 @@ from typing import Any, Callable, Dict
 import tomodachi
 from tomodachi import amqp, amqp_publish
 from tomodachi.discovery import DummyRegistry
-from tomodachi.protocol import JsonBase
+from tomodachi.envelope import JsonBase
 
 
-async def middleware_function(func: Callable, service: Any, message: Any, routing_key: str, context: Dict, *args: Any, **kwargs: Any) -> Any:
+async def middleware_function(
+    func: Callable, service: Any, message: Any, routing_key: str, context: Dict, *args: Any, **kwargs: Any
+) -> Any:
     # Functionality before function is called
-    service.log('middleware before')
+    service.log("middleware before")
 
     return_value = await func(*args, **kwargs)
 
@@ -17,37 +19,32 @@ async def middleware_function(func: Callable, service: Any, message: Any, routin
     # return_value = await func(*args, id='overridden', **kwargs)
 
     # Functinoality after function is called
-    service.log('middleware after')
+    service.log("middleware after")
 
     return return_value
 
 
-@tomodachi.service
 class ExampleAmqpService(tomodachi.Service):
-    name = 'example_amqp_service'
-    log_level = 'INFO'
-    uuid = os.environ.get('SERVICE_UUID')
+    name = "example-amqp-service"
+    log_level = "INFO"
+    uuid = os.environ.get("SERVICE_UUID")
 
     # Build own "discovery" functions, to be run on start and stop
     # See tomodachi/discovery/dummy_registry.py for example
     discovery = [DummyRegistry]
 
-    # The message protocol class defines how a message should be processed when sent and received
-    # See tomodachi/protocol/json_base.py for a basic example using JSON and transferring some metadata
-    message_protocol = JsonBase
+    # The message envelope class defines how a message should be processed when sent and received
+    # See tomodachi/envelope/json_base.py for a basic example using JSON and transferring some metadata
+    message_envelope = JsonBase
 
     # Adds a middleware function that is run on every incoming message.
     # Several middlewares can be chained.
     message_middleware = [middleware_function]
 
     # Some options can be specified to define credentials, used ports, hostnames, access log, etc.
-    options = {
-        'amqp': {
-            'queue_ttl': 60
-        }
-    }
+    options = {"amqp": {"queue_ttl": 60}}
 
-    @amqp('example.route1')
+    @amqp("example.route1")
     async def route1a(self, data: Any) -> None:
         self.log('Received data (function: route1a) - "{}"'.format(data))
 
@@ -56,4 +53,4 @@ class ExampleAmqpService(tomodachi.Service):
             self.log('Publish data "{}"'.format(data))
             await amqp_publish(self, data, routing_key=routing_key)
 
-        await publish('友達', 'example.route1')
+        await publish("友達", "example.route1")