+++ tests/services/aws_sns_sqs_service_with_credentials.py
@@ -6,7 +6,7 @@ from typing import Any
 
 import tomodachi
 from tomodachi.discovery.aws_sns_registration import AWSSNSRegistration
-from tomodachi.protocol.json_base import JsonBase
+from tomodachi.envelope.json_base import JsonBase
 from tomodachi.transport.aws_sns_sqs import aws_sns_sqs, aws_sns_sqs_publish
 
 data_uuid = str(uuid.uuid4())
@@ -17,7 +17,7 @@ class AWSSNSSQSService(tomodachi.Service):
     name = "test_aws_sns_sqs"
     log_level = "INFO"
     discovery = [AWSSNSRegistration]
-    message_protocol = JsonBase
+    message_envelope = JsonBase
     options = {
         "aws": {
             "region_name": os.environ.get("TOMODACHI_TEST_AWS_REGION"),