+++ tomodachi/protocol/__init__.py
@@ -1,15 +1,5 @@
-from typing import Any
-
-from tomodachi.protocol.json_base import JsonBase
-
-try:
-    from tomodachi.protocol.protobuf_base import ProtobufBase  # type: ignore
-except Exception:  # pragma: no cover
-
-    class ProtobufBase(object):  # type: ignore
-        @classmethod
-        def validate(cls, **kwargs: Any) -> None:
-            raise Exception("google.protobuf package not installed")
+from tomodachi.envelope.json_base import JsonBase
+from tomodachi.envelope.protobuf_base import ProtobufBase
 
 
 __all__ = ["JsonBase", "ProtobufBase"]